Consolidate pki-run-commands with pki-workflow-helpers
