+++ ui/tests/acceptance/pki/pki-configuration-test.js
@@ -14,7 +14,7 @@ import authPage from 'vault/tests/pages/auth';
 import logout from 'vault/tests/pages/logout';
 import enablePage from 'vault/tests/pages/settings/mount-secret-backend';
 import { runCmd } from 'vault/tests/helpers/commands';
-import { PKI_WORKFLOW } from 'vault/tests/helpers/pki/workflow';
+import { PKI_WORKFLOW } from 'vault/tests/helpers/pki/pki-workflow-helpers';
 import { issuerPemBundle } from 'vault/tests/helpers/pki/values';
 import { GENERAL } from 'vault/tests/helpers/general-selectors';
 import { PKI_GENERATE_ROOT } from 'vault/tests/helpers/components/pki/pki-generate-root';
